Decouple non-scene components from scenes
- Header is not a scene but a layout fragment. Its wirings are
  moved to its parents for now but they will eventually be
  abstracted in a layout component.

- Notification is simply a component so it is moved among
  components.

- Modal was complicated. First of all; the actual scene components
  we use in the modals (SignIn, SignUp, ..) are decoupled from the
  modal, as scenes should be agnostic of their containers, be it a
  page, a modal, or whatever. It's not their business.

  And the modal "scene" itself is actually a "modal router"
  component which routes to the correct modal scene according to
  the given modal name. So it is moved among components as well,
  with the name "ModalRouter".
