+++ src/scenes/Image/index.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import Img from './Img'
+
+const Image = ({ image: { src, title } }) => (
+    <Img src={src} alt={title} />
+)
+
+export default Image
