+++ src/scenes/PasswordReset/index.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import PasswordResetForm from 'containers/Forms/PasswordReset'
+
+const PasswordReset = () => <PasswordResetForm />
+
+export default PasswordReset
