+++ src/scenes/SignIn/index.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import SignInForm from 'containers/Forms/SignIn'
+
+const SignIn = () => <SignInForm />
+
+export default SignIn
