+++ src/scenes/SignUp/index.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import SignUpForm from 'containers/Forms/SignUp'
+
+const SignUp = () => <SignUpForm />
+
+export default SignUp
